pyproject.toml
[project]
name = "core-bff"
version = "0.1.0"
description = ""
authors = [
    {name = "Alejandro Morales",email = "ja.moralesv@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pyodbc (>=5.3.0,<6.0.0)",
    "azure-identity (>=1.25.2,<2.0.0)",
    "azure-keyvault-secrets (>=4.10.0,<5.0.0)",
    "aioodbc (>=0.5.0,<0.6.0)",
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
mypy = "^1.8.0"
ruff = "^0.3.0"
httpx = "^0.28.1"
pre-commit = "^4.5.1"

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "I"] # Errores, Pyflakes e Isort

[tool.mypy]
strict = true
ignore_missing_imports = true
plugins = [
  "pydantic.mypy"
]

# Esto ayuda a Mypy a entender que pydantic-settings tiene tipos
[[tool.mypy.overrides]]
module = ["pydantic_settings.*", "fastapi.*"]
ignore_missing_imports = false
